Add files via upload
Write a MIPS assembly program that
does the following:           
(i) Ask the user to enter the number of rows R and read it.           
(ii) Ask the user to enter the number of columns C and read it.           
(iii) Ask the user to enter a two-dimensional array of RxC characters. Elements of a single row should be separated by a single space and each row is read in a new line.           
(iv) Print the entered array in a new line (prints its entered dimensions as well).           
(v) Ask the user to enter two row numbers           
(vi) Exchange the two entered rows and print the array after the exchange.